support for device name
