+++ azure-communication-ui/azure-communication-ui/src/calling/java/com/azure/android/communication/ui/calling/redux/middleware/bluetooth/AudioSwitchMiddleware.kt
@@ -50,7 +50,6 @@ internal class AudioSwitchingMiddleware(
                                     AudioDeviceSelectionStatus.SPEAKER_SELECTED
                                 )
                             )
-
                     }
                     next(action)
